gradient_descent/requirements.txt
attrs==22.1.0
autograd==1.5
contourpy==1.0.5
cycler==0.11.0
flake8==5.0.4
fonttools==4.37.4
future==0.18.2
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.1
mccabe==0.7.0
numpy==1.23.4
packaging==21.3
Pillow==9.2.0
pkg_resources==0.0.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.1.3
python-dateutil==2.8.2
six==1.16.0
tomli==2.0.1
